--- deduce/exception.py
@@ -1,3 +1,0 @@
-class NestedTagsError(Exception):
-    def __init__(self, msg: str):
-        super().__init__(str)
